saving before seed restructure merge
