--- client/models/index.js
@@ -1,7 +1,0 @@
-// Exporting an object containing all of our models
-
-module.exports = {
-    User: require("./user"),
-    Activity: require("./activity")
-  };
-
