+++ models/index.js
@@ -1,3 +1,7 @@
+// Exporting an object containing all of our models
+
 module.exports = {
-  Book: require("./book")
-};
+    User: require("./user"),
+    Activity: require("./activity")
+  };
+
